Moved constants back to constants file
